requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyQt6>=6.5.0
umap-learn>=0.5.3
scipy>=1.10.0
flowsom>=0.2.2
anndata>=0.12.6
pillow>=11.0.0

# Optional dependencies
# phenograph (may require platform-specific installation)
# phenograph
